Added basic security

+++ api/routes/profiles.js
@@ -1,13 +1,10 @@
 const express = require('express');
 const Profiles = require('../models/profiles');
 
+import {getAuthenticatedUserName, authenticatedUserHasRole} from '../utils/SecurityHelper';
+
 export let router = express.Router();
 export let publicRouter = express.Router();
-
-const randomUuid = () => {
-    return Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15);
-}
-
 router.route("/")
     .get(async (request, response) => {
         let profiles = await Profiles.find({}, null).exec();
@@ -25,13 +22,28 @@ router.route("/")
 router.route("/:username")
     .get(async ({params: {username}}, response) => {
         let profile = await Profiles.findOne({username}, null).exec();
+        if (!profile) {
+            response.status(404);
+            return response.send();
+        }
+
         return response.json(profile);
     })
     .put(async ({params: {username}}, response) => {
+        if (getAuthenticatedUserName() !== username) {
+            response.status(403);
+            return response.send();
+        }
+
         let profile = await Profiles.updateOne({username}, profile);
         return response.json(profile);
     })
     .delete(async ({params: {username}}, response) => {
+        if (getAuthenticatedUserName() !== username) {
+            response.status(403);
+            return response.send();
+        }
+
         await Profiles.deleteOne({params: {username}});
         return response.send();
     })
@@ -39,5 +51,11 @@ router.route("/:username")
 publicRouter.route("/:id")
     .get(async ({params: {id: _id}}, response) => {
         let profile = await Profiles.findOne({_id}, null).exec();
+
+        if (!profile) {
+            response.status(404);
+            return response.send();
+        }
+
         return response.json(profile);
     })
